backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
PyJWT
python-multipart
requests
pydantic-settings 